Move to durationText determine
